refactor(ethereum): layer client, registry, adapters, and helpers
Restructure the ethereum provider to remove cycle dependencies and clarify
boundaries: client as gateway, registry for routing, standard/generic adapters
for contract logic, and helpers for shared RPC/pagination operations. Inject
BlockProvider directly, route GetTokenEvents through adapters with chain and
timestamp enrichment, and simplify error classification in core_executor.
