+++ internal/providers/ethereum/helpers/abi.go
@@ -1,4 +1,4 @@
-package adapter
+package helpers
 
 import (
 	"fmt"
