+++ internal/providers/ethereum/registry/config.go
@@ -1,4 +1,4 @@
-package adapter
+package registry
 
 import (
 	"encoding/json"
@@ -10,7 +10,10 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 
 	ethadapter "github.com/feral-file/ff-indexer-v2/internal/adapter"
+	"github.com/feral-file/ff-indexer-v2/internal/block"
 	"github.com/feral-file/ff-indexer-v2/internal/domain"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/adapters"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/helpers"
 )
 
 var eventSignaturePattern = regexp.MustCompile(`^[A-Za-z_][A-Za-z0-9_]*\([^)]+\)$`)
@@ -33,28 +36,11 @@ type ContractConfig struct {
 
 // AdapterConfig holds declarative method routing for a contract.
 type AdapterConfig struct {
-	Existence MethodConfig   `json:"existence"`
-	Owner     MethodConfig   `json:"owner"`
-	Metadata  MetadataConfig `json:"metadata"`
-	Events    []EventConfig  `json:"events,omitempty"`
+	Existence MethodConfig           `json:"existence"`
+	Owner     MethodConfig           `json:"owner"`
+	Metadata  MetadataConfig         `json:"metadata"`
+	Events    []adapters.EventConfig `json:"events,omitempty"`
 }
-
-// EventConfig describes a custom event for provenance indexing.
-type EventConfig struct {
-	Signature          string            `json:"signature"`
-	MapToStandardEvent domain.EventType  `json:"mapToStandardEvent"`
-	IndexedParams      []string          `json:"indexedParams"`
-	DataParams         []string          `json:"dataParams"`
-	ParameterMappings  map[string]string `json:"parameterMappings"`
-}
-
-// Supported event field names for parameterMappings values.
-const (
-	EventFieldFromAddress = "FromAddress"
-	EventFieldToAddress   = "ToAddress"
-	EventFieldTokenNumber = "TokenNumber"
-	EventFieldQuantity    = "Quantity"
-)
 
 // MethodConfig describes a contract method call used by GenericAdapter.
 type MethodConfig struct {
@@ -72,33 +58,15 @@ type MetadataConfig struct {
 
 // LoadContractsConfig reads and validates contracts.json from the provided filesystem.
 func LoadContractsConfig(fsys fs.FS) (*ContractsConfig, error) {
-	content, err := fs.ReadFile(fsys, contractsFile)
-	if err != nil {
-		return nil, fmt.Errorf("read contracts config: %w", err)
-	}
-
-	var cfg ContractsConfig
-	if err := json.Unmarshal(content, &cfg); err != nil {
-		return nil, fmt.Errorf("parse contracts config: %w", err)
-	}
-
-	// Load ABI registry for validation
-	abiRegistry, err := NewABIRegistry(fsys)
+	abiRegistry, err := helpers.NewABIRegistry(fsys)
 	if err != nil {
 		return nil, fmt.Errorf("load ABI registry for validation: %w", err)
 	}
-
-	if err := validateContractsConfig(&cfg, abiRegistry); err != nil {
-		return nil, err
-	}
-
-	return &cfg, nil
+	return LoadContractsConfigWithABIRegistry(fsys, abiRegistry)
 }
 
 // LoadContractsConfigWithABIRegistry reads and validates contracts.json using a pre-loaded ABI registry.
-//
-// Reason: Avoid loading the ABI registry twice during registry initialization.
-func LoadContractsConfigWithABIRegistry(fsys fs.FS, abiRegistry *ABIRegistry) (*ContractsConfig, error) {
+func LoadContractsConfigWithABIRegistry(fsys fs.FS, abiRegistry *helpers.ABIRegistry) (*ContractsConfig, error) {
 	content, err := fs.ReadFile(fsys, contractsFile)
 	if err != nil {
 		return nil, fmt.Errorf("read contracts config: %w", err)
@@ -116,7 +84,7 @@ func LoadContractsConfigWithABIRegistry(fsys fs.FS, abiRegistry *ABIRegistry) (*
 	return &cfg, nil
 }
 
-func validateContractsConfig(cfg *ContractsConfig, abiRegistry *ABIRegistry) error {
+func validateContractsConfig(cfg *ContractsConfig, abiRegistry *helpers.ABIRegistry) error {
 	seen := make(map[string]struct{}, len(cfg.Contracts))
 
 	for i := range cfg.Contracts {
@@ -135,7 +103,7 @@ func validateContractsConfig(cfg *ContractsConfig, abiRegistry *ABIRegistry) err
 	return nil
 }
 
-func validateContractEntry(entry *ContractConfig, abiRegistry *ABIRegistry) error {
+func validateContractEntry(entry *ContractConfig, abiRegistry *helpers.ABIRegistry) error {
 	if entry.Chain == "" {
 		return fmt.Errorf("chain is required")
 	}
@@ -161,17 +129,12 @@ func validateContractEntry(entry *ContractConfig, abiRegistry *ABIRegistry) erro
 		return fmt.Errorf("adapter.owner.abi is required")
 	}
 
-	// Validate existence method configuration
 	if err := validateMethodConfig(entry.Adapter.Existence, abiRegistry, "adapter.existence"); err != nil {
 		return err
 	}
-
-	// Validate owner method configuration
 	if err := validateMethodConfig(entry.Adapter.Owner, abiRegistry, "adapter.owner"); err != nil {
 		return err
 	}
-
-	// Validate metadata configuration
 	if err := validateMetadataConfig(entry.Adapter.Metadata, abiRegistry); err != nil {
 		return err
 	}
@@ -190,7 +153,7 @@ func validateContractEntry(entry *ContractConfig, abiRegistry *ABIRegistry) erro
 	return nil
 }
 
-func validateEventConfig(cfg EventConfig, fieldPath string) error {
+func validateEventConfig(cfg adapters.EventConfig, fieldPath string) error {
 	if cfg.Signature == "" {
 		return fmt.Errorf("%s: signature is required", fieldPath)
 	}
@@ -223,13 +186,11 @@ func validateEventConfig(cfg EventConfig, fieldPath string) error {
 		if name == "" {
 			return fmt.Errorf("%s: empty parameter name in dataParams", fieldPath)
 		}
-		// Check for duplicates within dataParams FIRST (before checking paramNames)
 		for j := 0; j < i; j++ {
 			if cfg.DataParams[j] == name {
 				return fmt.Errorf("%s: duplicate parameter name %q in dataParams", fieldPath, name)
 			}
 		}
-		// Then check for duplicates across indexed/data
 		if _, existsInIndexed := paramNames[name]; existsInIndexed {
 			return fmt.Errorf("%s: duplicate parameter name %q across indexedParams and dataParams", fieldPath, name)
 		}
@@ -242,7 +203,7 @@ func validateEventConfig(cfg EventConfig, fieldPath string) error {
 			return fmt.Errorf("%s: parameterMappings references unknown parameter %q", fieldPath, paramName)
 		}
 		switch eventField {
-		case EventFieldFromAddress, EventFieldToAddress, EventFieldTokenNumber, EventFieldQuantity:
+		case adapters.EventFieldFromAddress, adapters.EventFieldToAddress, adapters.EventFieldTokenNumber, adapters.EventFieldQuantity:
 		default:
 			return fmt.Errorf("%s: invalid parameterMappings target %q for parameter %q (allowed: FromAddress, ToAddress, TokenNumber, Quantity)", fieldPath, eventField, paramName)
 		}
@@ -259,17 +220,13 @@ func validateEventConfig(cfg EventConfig, fieldPath string) error {
 		}
 	}
 
-	if err := validateSemanticRequirements(cfg, usedTargetFields, fieldPath); err != nil {
-		return err
-	}
-
-	return nil
+	return validateSemanticRequirements(cfg, usedTargetFields, fieldPath)
 }
 
-func validateSemanticRequirements(cfg EventConfig, mappedFields map[string]string, fieldPath string) error {
-	hasTokenNumber := mappedFields[EventFieldTokenNumber] != ""
-	hasToAddress := mappedFields[EventFieldToAddress] != ""
-	hasFromAddress := mappedFields[EventFieldFromAddress] != ""
+func validateSemanticRequirements(cfg adapters.EventConfig, mappedFields map[string]string, fieldPath string) error {
+	hasTokenNumber := mappedFields[adapters.EventFieldTokenNumber] != ""
+	hasToAddress := mappedFields[adapters.EventFieldToAddress] != ""
+	hasFromAddress := mappedFields[adapters.EventFieldFromAddress] != ""
 
 	switch cfg.MapToStandardEvent {
 	case domain.EventTypeTransfer:
@@ -305,8 +262,7 @@ func validateSemanticRequirements(cfg EventConfig, mappedFields map[string]strin
 	return nil
 }
 
-// validateMethodConfig validates that a method exists in its declared ABI and has valid success condition.
-func validateMethodConfig(cfg MethodConfig, abiRegistry *ABIRegistry, fieldPath string) error {
+func validateMethodConfig(cfg MethodConfig, abiRegistry *helpers.ABIRegistry, fieldPath string) error {
 	contractABI, err := abiRegistry.Get(cfg.ABI)
 	if err != nil {
 		return fmt.Errorf("%s: ABI %q not found: %w", fieldPath, cfg.ABI, err)
@@ -316,11 +272,9 @@ func validateMethodConfig(cfg MethodConfig, abiRegistry *ABIRegistry, fieldPath 
 		return fmt.Errorf("%s: method %q not found in ABI %q", fieldPath, cfg.Method, cfg.ABI)
 	}
 
-	// Validate success condition if specified
 	if cfg.SuccessCondition != "" {
-		switch SuccessCondition(cfg.SuccessCondition) {
-		case SuccessNoRevert, SuccessAddressNonZero:
-			// Valid
+		switch adapters.SuccessCondition(cfg.SuccessCondition) {
+		case adapters.SuccessNoRevert, adapters.SuccessAddressNonZero:
 		default:
 			return fmt.Errorf("%s: invalid success_condition %q (allowed: no_revert, address_nonzero)", fieldPath, cfg.SuccessCondition)
 		}
@@ -329,19 +283,16 @@ func validateMethodConfig(cfg MethodConfig, abiRegistry *ABIRegistry, fieldPath 
 	return nil
 }
 
-// validateMetadataConfig validates metadata source and method configuration.
-func validateMetadataConfig(cfg MetadataConfig, abiRegistry *ABIRegistry) error {
-	// Validate metadata source
+func validateMetadataConfig(cfg MetadataConfig, abiRegistry *helpers.ABIRegistry) error {
 	if cfg.Source != "" {
-		switch MetadataSource(cfg.Source) {
-		case MetadataSourceOnChain, MetadataSourceVendorOnly:
-			// Valid
+		switch adapters.MetadataSource(cfg.Source) {
+		case adapters.MetadataSourceOnChain, adapters.MetadataSourceVendorOnly:
 		default:
 			return fmt.Errorf("adapter.metadata.source: invalid value %q (allowed: on_chain, vendor_only)", cfg.Source)
 		}
 	}
 
-	if MetadataSource(cfg.Source) == MetadataSourceOnChain {
+	if adapters.MetadataSource(cfg.Source) == adapters.MetadataSourceOnChain {
 		if cfg.Method == nil {
 			return fmt.Errorf("adapter.metadata.method is required when source is on_chain")
 		}
@@ -357,31 +308,31 @@ func contractKey(chain domain.Chain, address string) string {
 	return fmt.Sprintf("%s:%s", chain, strings.ToLower(common.HexToAddress(address).Hex()))
 }
 
-func buildMethodCall(cfg MethodConfig, abiRegistry *ABIRegistry) (*MethodCall, error) {
+func buildMethodCall(cfg MethodConfig, abiRegistry *helpers.ABIRegistry) (*adapters.MethodCall, error) {
 	contractABI, err := abiRegistry.Get(cfg.ABI)
 	if err != nil {
 		return nil, err
 	}
 
-	return &MethodCall{
+	return &adapters.MethodCall{
 		Method:           cfg.Method,
 		ABI:              contractABI,
 		Params:           cfg.Params,
-		SuccessCondition: SuccessCondition(cfg.SuccessCondition),
+		SuccessCondition: adapters.SuccessCondition(cfg.SuccessCondition),
 	}, nil
 }
 
-func buildMetadataConfig(cfg MetadataConfig, abiRegistry *ABIRegistry) (ContractMetadataConfig, error) {
-	source := MetadataSource(cfg.Source)
+func buildMetadataConfig(cfg MetadataConfig, abiRegistry *helpers.ABIRegistry) (adapters.ContractMetadataConfig, error) {
+	source := adapters.MetadataSource(cfg.Source)
 	if source == "" {
-		source = MetadataSourceVendorOnly
+		source = adapters.MetadataSourceVendorOnly
 	}
 
-	metadata := ContractMetadataConfig{Source: source}
+	metadata := adapters.ContractMetadataConfig{Source: source}
 	if cfg.Method != nil {
 		method, err := buildMethodCall(*cfg.Method, abiRegistry)
 		if err != nil {
-			return ContractMetadataConfig{}, err
+			return adapters.ContractMetadataConfig{}, err
 		}
 		metadata.Method = method
 	}
@@ -392,9 +343,13 @@ func buildMetadataConfig(cfg MetadataConfig, abiRegistry *ABIRegistry) (Contract
 // BuildGenericAdapterFromConfig constructs a GenericAdapter for a validated contract entry.
 func BuildGenericAdapterFromConfig(
 	entry ContractConfig,
-	abiRegistry *ABIRegistry,
+	abiRegistry *helpers.ABIRegistry,
 	ethClient ethadapter.EthClient,
-) (ContractAdapter, error) {
+	pagination *helpers.PaginationHelper,
+	chainID domain.Chain,
+	blockProvider block.BlockProvider,
+	clock ethadapter.Clock,
+) (adapters.ContractAdapter, error) {
 	existence, err := buildMethodCall(entry.Adapter.Existence, abiRegistry)
 	if err != nil {
 		return nil, fmt.Errorf("existence method: %w", err)
@@ -412,13 +367,18 @@ func BuildGenericAdapterFromConfig(
 
 	supportsProvenance := len(entry.Adapter.Events) > 0
 
-	return NewGenericAdapter(
+	return adapters.NewGenericAdapter(
+		entry.Standard,
 		existence,
 		owner,
 		metadata,
 		ethClient,
+		pagination,
 		supportsProvenance,
 		entry.Adapter.Events,
+		chainID,
+		blockProvider,
+		clock,
 	), nil
 }
 
@@ -430,10 +390,10 @@ func (r *AdapterRegistry) IsVendorOnlyMetadata(chain domain.Chain, contractAddre
 		return false
 	}
 
-	source := MetadataSource(entry.Adapter.Metadata.Source)
+	source := adapters.MetadataSource(entry.Adapter.Metadata.Source)
 	if source == "" {
 		return true
 	}
 
-	return source == MetadataSourceVendorOnly
+	return source == adapters.MetadataSourceVendorOnly
 }
