+++ internal/providers/ethereum/registry/registry_test.go
@@ -1,16 +1,33 @@
-package adapter_test
+package registry_test
 
 import (
 	"context"
+	"math/big"
 	"testing"
+	"testing/fstest"
 
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
 
+	ethadapter "github.com/feral-file/ff-indexer-v2/internal/adapter"
 	"github.com/feral-file/ff-indexer-v2/internal/domain"
 	"github.com/feral-file/ff-indexer-v2/internal/mocks"
-	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/adapter"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/adapters"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/helpers"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/registry"
 )
+
+const cryptopunksABI = `[
+  {
+    "constant": true,
+    "inputs": [{"name": "index", "type": "uint256"}],
+    "name": "punkIndexToAddress",
+    "outputs": [{"name": "", "type": "address"}],
+    "type": "function"
+  }
+]`
 
 const cryptopunksContractConfigWithEvents = `{
   "contracts": [
@@ -89,21 +106,40 @@ const cryptopunksContractConfig = `{
   ]
 }`
 
+func testContractFS(t *testing.T, contractsJSON string) fstest.MapFS {
+	t.Helper()
+
+	return fstest.MapFS{
+		"contracts.json":        {Data: []byte(contractsJSON)},
+		"abis/cryptopunks.json": {Data: []byte(cryptopunksABI)},
+	}
+}
+
+func newTestRegistry(t *testing.T, mockClient *mocks.MockEthClient, fs fstest.MapFS) *registry.AdapterRegistry {
+	t.Helper()
+
+	pagination := helpers.NewPaginationHelper(mockClient, ethadapter.NewClock(), nil)
+	reg, err := registry.NewAdapterRegistry(
+		fs,
+		mockClient,
+		ethadapter.NewClock(),
+		nil,
+		pagination,
+		domain.ChainEthereumMainnet,
+	)
+	require.NoError(t, err)
+	return reg
+}
+
 func TestAdapterRegistry_GetAdapter(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockClient := mocks.NewMockEthClient(ctrl)
-	ops := &stubStandardOps{owner: "0x1234567890123456789012345678901234567890"}
 
-	registry, err := adapter.NewAdapterRegistry(
-		testContractFS(t, cryptopunksContractConfig),
-		mockClient,
-		ops,
-	)
-	require.NoError(t, err)
-	require.Equal(t, 1, registry.ContractOverrideCount())
+	reg := newTestRegistry(t, mockClient, testContractFS(t, cryptopunksContractConfig))
+	require.Equal(t, 1, reg.ContractOverrideCount())
 
 	t.Run("configured contract uses generic adapter", func(t *testing.T) {
-		adp, err := registry.GetAdapter(
+		adp, err := reg.GetAdapter(
 			domain.ChainEthereumMainnet,
 			"0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb",
 			domain.StandardERC721,
@@ -113,7 +149,12 @@ func TestAdapterRegistry_GetAdapter(t *testing.T) {
 	})
 
 	t.Run("unknown contract uses standard adapter", func(t *testing.T) {
-		adp, err := registry.GetAdapter(
+		ownerAddr := common.HexToAddress("0x1234567890123456789012345678901234567890")
+		mockClient.EXPECT().
+			CallContract(gomock.Any(), gomock.Any(), gomock.Nil()).
+			Return(common.LeftPadBytes(ownerAddr.Bytes(), 32), nil)
+
+		adp, err := reg.GetAdapter(
 			domain.ChainEthereumMainnet,
 			"0x0000000000000000000000000000000000000001",
 			domain.StandardERC721,
@@ -124,24 +165,23 @@ func TestAdapterRegistry_GetAdapter(t *testing.T) {
 		exists, err := adp.TokenExists(context.Background(), "0x0000000000000000000000000000000000000001", "1")
 		require.NoError(t, err)
 		require.True(t, exists)
-		require.True(t, ops.ownerOfCalled)
 	})
 
 	t.Run("unsupported standard returns error", func(t *testing.T) {
-		_, err := registry.GetAdapter(
+		_, err := reg.GetAdapter(
 			domain.ChainEthereumMainnet,
 			"0x0000000000000000000000000000000000000001",
 			domain.StandardFA2,
 		)
-		require.ErrorIs(t, err, adapter.ErrUnsupportedContractStandard)
+		require.ErrorIs(t, err, adapters.ErrUnsupportedContractStandard)
 	})
 
 	t.Run("vendor only metadata flag", func(t *testing.T) {
-		require.True(t, registry.IsVendorOnlyMetadata(
+		require.True(t, reg.IsVendorOnlyMetadata(
 			domain.ChainEthereumMainnet,
 			"0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb",
 		))
-		require.False(t, registry.IsVendorOnlyMetadata(
+		require.False(t, reg.IsVendorOnlyMetadata(
 			domain.ChainEthereumMainnet,
 			"0x0000000000000000000000000000000000000001",
 		))
@@ -152,15 +192,10 @@ func TestAdapterRegistry_EmptyConfig(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockClient := mocks.NewMockEthClient(ctrl)
 
-	registry, err := adapter.NewAdapterRegistry(
-		testContractFS(t, `{"contracts": []}`),
-		mockClient,
-		&stubStandardOps{},
-	)
-	require.NoError(t, err)
-	require.Equal(t, 0, registry.ContractOverrideCount())
+	reg := newTestRegistry(t, mockClient, testContractFS(t, `{"contracts": []}`))
+	require.Equal(t, 0, reg.ContractOverrideCount())
 
-	adp, err := registry.GetAdapter(domain.ChainEthereumMainnet, "0xabc", domain.StandardERC1155)
+	adp, err := reg.GetAdapter(domain.ChainEthereumMainnet, "0xabc", domain.StandardERC1155)
 	require.NoError(t, err)
 	require.True(t, adp.SupportsProvenance())
 }
@@ -169,14 +204,9 @@ func TestAdapterRegistry_SupportsProvenance(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockClient := mocks.NewMockEthClient(ctrl)
 
-	registry, err := adapter.NewAdapterRegistry(
-		testContractFS(t, cryptopunksContractConfig),
-		mockClient,
-		&stubStandardOps{},
-	)
-	require.NoError(t, err)
+	reg := newTestRegistry(t, mockClient, testContractFS(t, cryptopunksContractConfig))
 
-	supported, err := registry.SupportsProvenance(
+	supported, err := reg.SupportsProvenance(
 		domain.ChainEthereumMainnet,
 		"0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb",
 		domain.StandardERC721,
@@ -184,7 +214,7 @@ func TestAdapterRegistry_SupportsProvenance(t *testing.T) {
 	require.NoError(t, err)
 	require.False(t, supported)
 
-	supported, err = registry.SupportsProvenance(
+	supported, err = reg.SupportsProvenance(
 		domain.ChainEthereumMainnet,
 		"0x0000000000000000000000000000000000000001",
 		domain.StandardERC721,
@@ -192,26 +222,21 @@ func TestAdapterRegistry_SupportsProvenance(t *testing.T) {
 	require.NoError(t, err)
 	require.True(t, supported)
 
-	_, err = registry.SupportsProvenance(
+	_, err = reg.SupportsProvenance(
 		domain.ChainEthereumMainnet,
 		"0x0000000000000000000000000000000000000001",
 		domain.StandardFA2,
 	)
-	require.ErrorIs(t, err, adapter.ErrUnsupportedContractStandard)
+	require.ErrorIs(t, err, adapters.ErrUnsupportedContractStandard)
 }
 
 func TestAdapterRegistry_SupportsProvenance_WithCustomEvents(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockClient := mocks.NewMockEthClient(ctrl)
 
-	registry, err := adapter.NewAdapterRegistry(
-		testContractFS(t, cryptopunksContractConfigWithEvents),
-		mockClient,
-		&stubStandardOps{},
-	)
-	require.NoError(t, err)
+	reg := newTestRegistry(t, mockClient, testContractFS(t, cryptopunksContractConfigWithEvents))
 
-	supported, err := registry.SupportsProvenance(
+	supported, err := reg.SupportsProvenance(
 		domain.ChainEthereumMainnet,
 		"0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb",
 		domain.StandardERC721,
@@ -219,26 +244,51 @@ func TestAdapterRegistry_SupportsProvenance_WithCustomEvents(t *testing.T) {
 	require.NoError(t, err)
 	require.True(t, supported)
 
-	adp, err := registry.GetAdapter(
+	adp, err := reg.GetAdapter(
 		domain.ChainEthereumMainnet,
 		"0xb47e3cd837ddf8e4c57f05d70ab865de6e193bbb",
 		domain.StandardERC721,
 	)
 	require.NoError(t, err)
-	require.Len(t, adp.GetProvenanceEventConfigs(), 2)
+	require.Len(t, adp.GetEventSignatures(), 2)
 }
 
 func TestAdapterRegistry_GetAllCustomEventSignatures(t *testing.T) {
 	ctrl := gomock.NewController(t)
 	mockClient := mocks.NewMockEthClient(ctrl)
 
-	registry, err := adapter.NewAdapterRegistry(
-		testContractFS(t, cryptopunksContractConfigWithEvents),
-		mockClient,
-		&stubStandardOps{},
-	)
-	require.NoError(t, err)
+	reg := newTestRegistry(t, mockClient, testContractFS(t, cryptopunksContractConfigWithEvents))
 
-	signatures := registry.GetAllCustomEventSignatures()
+	signatures := reg.GetAllCustomEventSignatures()
 	require.Len(t, signatures, 2)
+}
+
+func TestAdapterRegistry_ParseEvent_StandardERC721Transfer(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	mockClient := mocks.NewMockEthClient(ctrl)
+	reg := newTestRegistry(t, mockClient, testContractFS(t, `{"contracts": []}`))
+
+	from := common.HexToAddress("0x1111111111111111111111111111111111111111")
+	to := common.HexToAddress("0x2222222222222222222222222222222222222222")
+	tokenID := big.NewInt(42)
+
+	vLog := typesLog(from, to, tokenID)
+	event := &domain.BlockchainEvent{Chain: domain.ChainEthereumMainnet}
+	parsed, err := reg.ParseEvent(context.Background(), domain.ChainEthereumMainnet, vLog, event)
+	require.NoError(t, err)
+	require.Equal(t, domain.StandardERC721, parsed.Standard)
+	require.Equal(t, "42", parsed.TokenNumber)
+}
+
+// typesLog builds a minimal ERC721 Transfer log for tests.
+func typesLog(from, to common.Address, tokenID *big.Int) types.Log {
+	return types.Log{
+		Address: common.HexToAddress("0x0000000000000000000000000000000000000001"),
+		Topics: []common.Hash{
+			common.HexToHash("0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"),
+			common.BytesToHash(from.Bytes()),
+			common.BytesToHash(to.Bytes()),
+			common.BigToHash(tokenID),
+		},
+	}
 }
