+++ internal/providers/ethereum/subscriber.go
@@ -9,14 +9,14 @@ import (
 	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
 
 	"go.uber.org/zap"
 
 	"github.com/feral-file/ff-indexer-v2/internal/blockchain"
 	"github.com/feral-file/ff-indexer-v2/internal/domain"
 	"github.com/feral-file/ff-indexer-v2/internal/logger"
-	contractadapter "github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/adapter"
+	"github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/helpers"
+	contractregistry "github.com/feral-file/ff-indexer-v2/internal/providers/ethereum/registry"
 )
 
 const (
@@ -33,34 +33,11 @@ type ethSubscriber struct {
 	client          EthereumClient
 	chainID         domain.Chain
 	cfg             Config
-	adapterRegistry *contractadapter.AdapterRegistry
+	adapterRegistry *contractregistry.AdapterRegistry
 }
 
-// Event signatures
-var (
-	// Transfer event signature - shared by ERC20 and ERC721
-	// ERC20: Transfer(address indexed from, address indexed to, uint256 value) - 3 topics
-	// ERC721: Transfer(address indexed from, address indexed to, uint256 indexed tokenId) - 4 topics
-	transferEventSignature = crypto.Keccak256Hash([]byte("Transfer(address,address,uint256)"))
-
-	// ERC1155 TransferSingle(address indexed operator, address indexed from, address indexed to, uint256 id, uint256 value)
-	transferSingleEventSignature = crypto.Keccak256Hash([]byte("TransferSingle(address,address,address,uint256,uint256)"))
-
-	// ERC1155 TransferBatch(address indexed operator, address indexed from, address indexed to, uint256[] ids, uint256[] values)
-	transferBatchEventSignature = crypto.Keccak256Hash([]byte("TransferBatch(address,address,address,uint256[],uint256[])"))
-
-	// EIP-4906 MetadataUpdate(uint256 _tokenId)
-	metadataUpdateEventSignature = crypto.Keccak256Hash([]byte("MetadataUpdate(uint256)"))
-
-	// EIP-4906 BatchMetadataUpdate(uint256 _fromTokenId, uint256 _toTokenId)
-	batchMetadataUpdateEventSignature = crypto.Keccak256Hash([]byte("BatchMetadataUpdate(uint256,uint256)"))
-
-	// ERC1155 URI(string _value, uint256 indexed _id)
-	uriEventSignature = crypto.Keccak256Hash([]byte("URI(string,uint256)"))
-)
-
 // NewSubscriber creates a new Ethereum event subscriber.
-func NewSubscriber(cfg Config, ethereumClient EthereumClient, adapterRegistry *contractadapter.AdapterRegistry) (blockchain.EventSource, error) {
+func NewSubscriber(cfg Config, ethereumClient EthereumClient, adapterRegistry *contractregistry.AdapterRegistry) (blockchain.EventSource, error) {
 	return &ethSubscriber{
 		client:          ethereumClient,
 		chainID:         cfg.ChainID,
@@ -71,14 +48,7 @@ func NewSubscriber(cfg Config, ethereumClient EthereumClient, adapterRegistry *c
 
 // SubscribeEvents subscribes to ERC721/ERC1155 transfer and metadata update events
 func (s *ethSubscriber) SubscribeEvents(ctx context.Context, fromBlock uint64, handler blockchain.EventHandler) error {
-	allEventSignatures := []common.Hash{
-		transferEventSignature,            // ERC20/ERC721 Transfer (will filter ERC20 in parseLog)
-		transferSingleEventSignature,      // ERC1155 TransferSingle
-		transferBatchEventSignature,       // ERC1155 TransferBatch
-		metadataUpdateEventSignature,      // EIP-4906 MetadataUpdate
-		batchMetadataUpdateEventSignature, // EIP-4906 BatchMetadataUpdate
-		uriEventSignature,                 // ERC1155 URI
-	}
+	allEventSignatures := helpers.StandardEventSignatures()
 	if s.adapterRegistry != nil {
 		allEventSignatures = append(allEventSignatures, s.adapterRegistry.GetAllCustomEventSignatures()...)
 	}
